pyproject.toml
[project]
name = "receipt-printer"
version = "0.1.0"
description = "Epson TM-M30II Bluetooth receipt printer controller for Raspberry Pi"
requires-python = ">=3.11"
dependencies = [
    "python-escpos>=3.0",
    "pyserial>=3.5",
    "requests>=2.31",
    "feedparser>=6.0",
    "apscheduler>=3.10",
]

[project.scripts]
receipt-printer = "receipt_printer.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/receipt_printer"]

[tool.ruff]
line-length = 88
target-version = "py311"
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = []
